+++ R/zzz.R
@@ -1,3 +1,4 @@
 .onLoad <- function(libname, pkgname) {
     set_praat_path(default.path = TRUE)
+    set_lmt_path(default.path = TRUE)
 }